Porting and updating code for API changes.
