+++ generate-dev-sln.ps1
@@ -3,13 +3,20 @@ param
 (
     [ValidateNotNullOrEmpty()]
     [string]
-    $DiagRepoRoot = "$PSScriptRoot\..\diagnostics"
+    $DiagRepoRoot = "$PSScriptRoot\..\diagnostics",
+
+    [ValidateNotNullOrEmpty()]
+    [string]
+    $OtelRepoRoot = "$PSScriptRoot\..\opentelemetry-dotnet-instrumentation"
 )
 
 $ErrorActionPreference = 'Stop'
 
 $resolvedPath = Resolve-Path $DiagRepoRoot
 $env:DIAGNOSTICS_REPO_ROOT=$resolvedPath
+
+$resolvedPath = Resolve-Path $OtelRepoRoot
+$env:OTEL_REPO_ROOT=$resolvedPath
 
 #Generates a solution that spans both the diagnostics and the dotnet-monitor repo.
 #This can be used to build both projects in VS.
@@ -29,6 +36,15 @@ if ($LASTEXITCODE -gt 0)
 }
 & dotnet sln $devSln add "$env:DIAGNOSTICS_REPO_ROOT\src\Microsoft.Diagnostics.NETCore.Client\Microsoft.Diagnostics.NETCore.Client.csproj"
 if ($LASTEXITCODE -gt 0)
+{
+    exit $LASTEXITCODE
+}
+& dotnet sln $devSln add "$env:OTEL_REPO_ROOT\next-gen\src\OpenTelemetry.AutoInstrumentation.Sdk\OpenTelemetry.AutoInstrumentation.Sdk.csproj"
+if ($LASTEXITCODE -gt 0)
+{
+    exit $LASTEXITCODE
+}
+& dotnet sln $devSln add "$env:OTEL_REPO_ROOT\next-gen\src\OpenTelemetry.AutoInstrumentation.Sdk.Configuration\OpenTelemetry.AutoInstrumentation.Sdk.Configuration.csproj"if ($LASTEXITCODE -gt 0)
 {
     exit $LASTEXITCODE
 }
